Push sampling logic into separate package
