+++ inference/core/active_learning/samplers/close_to_threshold.py
@@ -12,26 +12,6 @@ from inference.core.active_learning.entities import (
 from inference.core.constants import CLASSIFICATION_TASK
 
 
-def initialize_random_sampling(strategy_config: Dict[str, Any]) -> SamplingMethod:
-    sample_function = partial(
-        sample_randomly,
-        traffic_percentage=strategy_config["traffic_percentage"],
-    )
-    return SamplingMethod(
-        name=strategy_config["name"],
-        sample=sample_function,
-    )
-
-
-def sample_randomly(
-    image: np.ndarray,
-    prediction: Prediction,
-    prediction_type: PredictionType,
-    traffic_percentage: float,
-) -> bool:
-    return random.random() < traffic_percentage
-
-
 def initialize_close_to_threshold_sampling(
     strategy_config: Dict[str, Any]
 ) -> SamplingMethod:
